# Assignment 12/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelPay Fraud Detection API",
    "description": "Real-time fraud detection and prevention REST API. Implements the Transaction ingestion pipeline (FR-01\u2013FR-08), Fraud Case management (FR-09\u2013FR-10), and ML Model lifecycle (FR-13\u2013FR-14) from the SentinelPay System Requirements Document.\n\n**Author:** Teboho Mokoni  \n**Assignment:** 12 \u2013 Service Layer and REST API  \n**Stack:** Python 3.12 \u00b7 FastAPI \u00b7 In-memory repositories (A11)",
    "contact": {
      "name": "Teboho Mokoni",
      "url": "https://github.com/Teboho66/SentinelPay"
    },
    "license": {
      "name": "Academic project \u2014 CPUT Postgraduate Diploma"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/ml-models": {
      "post": {
        "tags": [
          "ML Models"
        ],
        "summary": "Register a new ML model version",
        "description": "FR-13: Registers a new model version in TRAINING stage. model_name must be one of: XGBOOST, ISOLATION_FOREST, DISTILBERT. Model ID is auto-generated as '{model_name_lower}-{version}'.",
        "operationId": "register_model_api_ml_models_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterModelRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Model registered in TRAINING stage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MLModelResponse"
                }
              }
            }
          },
          "409": {
            "description": "Model version already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid model name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "ML Models"
        ],
        "summary": "Get all registered ML models",
        "description": "Returns all model versions. Filter by stage or model_name using query parameters.",
        "operationId": "get_all_models_api_ml_models_get",
        "parameters": [
          {
            "name": "stage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Stage"
            },
            "example": "PRODUCTION"
          },
          {
            "name": "model_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Model Name"
            },
            "example": "XGBOOST"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MLModelResponse"
                  },
                  "title": "Response Get All Models Api Ml Models Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ml-models/{model_id}": {
      "get": {
        "tags": [
          "ML Models"
        ],
        "summary": "Get a model by ID",
        "operationId": "get_model_api_ml_models__model_id__get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MLModelResponse"
                }
              }
            }
          },
          "404": {
            "description": "Model not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ML Models"
        ],
        "summary": "Delete a model version",
        "description": "PRODUCTION models cannot be deleted. Promote a replacement first.",
        "operationId": "delete_model_api_ml_models__model_id__delete",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Model not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Cannot delete PRODUCTION model",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ml-models/{model_id}/evaluate": {
      "post": {
        "tags": [
          "ML Models"
        ],
        "summary": "Record evaluation metrics for a model",
        "description": "FR-13: Records precision, recall, F1, and AUC-ROC from holdout-set evaluation. Returns meets_promotion_gate: true/false indicating readiness for PRODUCTION promotion. Gate: precision \u2265 0.85 AND recall \u2265 0.80 (BR-ML1).",
        "operationId": "evaluate_model_api_ml_models__model_id__evaluate_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateModelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Metrics recorded; meets_promotion_gate indicates readiness",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluateModelResponse"
                }
              }
            }
          },
          "404": {
            "description": "Model not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Model not in TRAINING or STAGING stage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ml-models/{model_id}/promote": {
      "patch": {
        "tags": [
          "ML Models"
        ],
        "summary": "Promote a model to the next lifecycle stage",
        "description": "FR-13: Advances the model through TRAINING \u2192 STAGING \u2192 PRODUCTION \u2192 ARCHIVED. Promoting to PRODUCTION enforces BR-ML1 (precision \u2265 0.85 AND recall \u2265 0.80). Existing PRODUCTION model of the same type is automatically ARCHIVED (FR-13).",
        "operationId": "promote_model_api_ml_models__model_id__promote_patch",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoteModelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model promoted to target stage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MLModelResponse"
                }
              }
            }
          },
          "404": {
            "description": "Model not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Invalid stage transition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Promotion gate failed (BR-ML1)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ml-models/{model_id}/hot-swap": {
      "patch": {
        "tags": [
          "ML Models"
        ],
        "summary": "Hot-swap model artifact in PRODUCTION",
        "description": "FR-14: Reloads the model artifact for a PRODUCTION model without service restart. Only callable on PRODUCTION-stage models.",
        "operationId": "hot_swap_api_ml_models__model_id__hot_swap_patch",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotSwapRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Artifact path updated; hot-swap complete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MLModelResponse"
                }
              }
            }
          },
          "404": {
            "description": "Model not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Model must be in PRODUCTION stage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code"
          }
        },
        "type": "object",
        "required": [
          "error",
          "detail",
          "status_code"
        ],
        "title": "ErrorResponse"
      },
      "EvaluateModelRequest": {
        "properties": {
          "precision": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Precision",
            "example": 0.91
          },
          "recall": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Recall",
            "example": 0.87
          },
          "f1_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "F1 Score",
            "example": 0.89
          },
          "auc_roc": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Auc Roc",
            "example": 0.94
          }
        },
        "type": "object",
        "required": [
          "precision",
          "recall",
          "f1_score",
          "auc_roc"
        ],
        "title": "EvaluateModelRequest",
        "description": "POST /api/ml-models/{id}/evaluate \u2014 record holdout-set evaluation metrics."
      },
      "EvaluateModelResponse": {
        "properties": {
          "model": {
            "$ref": "#/components/schemas/MLModelResponse"
          },
          "meets_promotion_gate": {
            "type": "boolean",
            "title": "Meets Promotion Gate"
          },
          "precision": {
            "type": "number",
            "title": "Precision"
          },
          "recall": {
            "type": "number",
            "title": "Recall"
          },
          "f1_score": {
            "type": "number",
            "title": "F1 Score"
          },
          "auc_roc": {
            "type": "number",
            "title": "Auc Roc"
          }
        },
        "type": "object",
        "required": [
          "model",
          "meets_promotion_gate",
          "precision",
          "recall",
          "f1_score",
          "auc_roc"
        ],
        "title": "EvaluateModelResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HotSwapRequest": {
        "properties": {
          "new_artifact_path": {
            "type": "string",
            "title": "New Artifact Path",
            "example": "mlflow://models/xgboost/3.2-hotfix"
          }
        },
        "type": "object",
        "required": [
          "new_artifact_path"
        ],
        "title": "HotSwapRequest",
        "description": "PATCH /api/ml-models/{id}/hot-swap \u2014 reload artifact in PRODUCTION without restart."
      },
      "MLModelResponse": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "precision": {
            "type": "number",
            "title": "Precision"
          },
          "recall": {
            "type": "number",
            "title": "Recall"
          },
          "f1_score": {
            "type": "number",
            "title": "F1 Score"
          },
          "auc_roc": {
            "type": "number",
            "title": "Auc Roc"
          },
          "artifact_path": {
            "type": "string",
            "title": "Artifact Path"
          },
          "feature_schema_version": {
            "type": "string",
            "title": "Feature Schema Version"
          },
          "meets_promotion_gate": {
            "type": "boolean",
            "title": "Meets Promotion Gate"
          },
          "trained_at": {
            "type": "string",
            "format": "date-time",
            "title": "Trained At"
          },
          "promoted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promoted At"
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "model_name",
          "version",
          "stage",
          "precision",
          "recall",
          "f1_score",
          "auc_roc",
          "artifact_path",
          "feature_schema_version",
          "meets_promotion_gate",
          "trained_at",
          "promoted_at"
        ],
        "title": "MLModelResponse",
        "description": "Response schema for MLModel."
      },
...
      "PromoteModelRequest": {
        "properties": {
          "target_stage": {
            "type": "string",
            "title": "Target Stage",
            "description": "One of: STAGING, PRODUCTION, ARCHIVED, REJECTED",
            "example": "STAGING"
          }
        },
        "type": "object",
        "required": [
          "target_stage"
        ],
        "title": "PromoteModelRequest",
        "description": "PATCH /api/ml-models/{id}/promote \u2014 advance model through lifecycle stages."
      },
      "RegisterModelRequest": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "description": "One of: XGBOOST, ISOLATION_FOREST, DISTILBERT",
            "example": "XGBOOST"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "example": "3.2"
          },
          "artifact_path": {
            "type": "string",
            "title": "Artifact Path",
            "example": "mlflow://models/xgboost/3.2"
          },
          "feature_schema_version": {
            "type": "string",
            "title": "Feature Schema Version",
            "example": "tabular-v3"
          }
        },
        "type": "object",
        "required": [
          "model_name",
          "version",
          "artifact_path",
          "feature_schema_version"
        ],
        "title": "RegisterModelRequest",
        "description": "POST /api/ml-models \u2014 register a new model version."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}